+++ src/neo4j_genai/types.py
@@ -201,6 +201,7 @@ class VectorRetrieverModel(BaseModel):
     index_name: str
     embedder_model: Optional[EmbedderModel] = None
     return_properties: Optional[list[str]] = None
+    neo4j_database: Optional[str] = None
 
 
 class VectorCypherRetrieverModel(BaseModel):
@@ -208,6 +209,7 @@ class VectorCypherRetrieverModel(BaseModel):
     index_name: str
     retrieval_query: str
     embedder_model: Optional[EmbedderModel] = None
+    neo4j_database: Optional[str] = None
 
 
 class HybridRetrieverModel(BaseModel):
@@ -216,6 +218,7 @@ class HybridRetrieverModel(BaseModel):
     fulltext_index_name: str
     embedder_model: Optional[EmbedderModel] = None
     return_properties: Optional[list[str]] = None
+    neo4j_database: Optional[str] = None
 
 
 class HybridCypherRetrieverModel(BaseModel):
@@ -224,6 +227,7 @@ class HybridCypherRetrieverModel(BaseModel):
     fulltext_index_name: str
     retrieval_query: str
     embedder_model: Optional[EmbedderModel] = None
+    neo4j_database: Optional[str] = None
 
 
 class Text2CypherRetrieverModel(BaseModel):